requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.2.0
django-import-export==0.5.1
django-reversion==2.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.9
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.10.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0